26.01.21 record 1. 完成了多组录制的代码camera_record.py 2. 录制前，更新一次相机拍摄相关参数。检查参数合理性。录制时，禁用参数输入。 3. 修正了_para.py中all_para_dict和gui2_para.py中all_para_settings变量类型不统一的问题。并把拍摄时长相关的变量类型从int转换为float。 4. PixelFormat从Mono8转换为和SensorBitDepth同步的格式(原本img会保存为Mono8)。显示界面保持8bit绘制（gui2_image.py中，_show_ndarray_on_canvas内调用_to_uint8_for_display），保证性能。 5. 增加了选择路径时file_chooser的title，选择路径时可以更好地提示。 6. 更新了_para.py中部分参数的默认值。

+++ utils/gui2/gui2_para.py
@@ -1,3 +1,5 @@
+# 修改 all_para_settings 中的变量类型时，需要在 _para.py 中同步修改 all_para_dict 中的变量类型
+
 all_para_settings = {
     "camera": {
         # "brief": [
@@ -11,9 +13,9 @@ all_para_settings = {
             {"str": "样本数量", "name": "sample_num", "type": "int"},
             {"str": "帧率 (fps)", "name": "vimbaX_AcquisitionFrameRate", "type": "float"},
             # {"str": "", "name": "range=(1.0369e-05, 289.04776)", "type": "label"},
-            {"str": "单次拍摄时长 (s) ", "name": "video_period", "type": "int"},
-            {"str": "拍摄时间间隔 (min)", "name": "video_interval", "type": "int"},
-            {"str": "拍摄截止时间 (min)", "name": "video_end_time", "type": "int"},
+            {"str": "单次拍摄时长 (s) ", "name": "video_period", "type": "float"},
+            {"str": "拍摄时间间隔 (min)", "name": "video_interval", "type": "float"},
+            {"str": "拍摄截止时间 (min)", "name": "video_end_time", "type": "float"},
 
             {"str": "— 相机参数 —", "name": "", "type": "label"},
             {"str": "曝光时间", "name": "vimbaX_ExposureTime", "type": "float"},
